--- src/defaultFields/storageForIFRS.js
@@ -1,12 +1,0 @@
-export let newIfrsSchema = {
-    "type": "object",
-    "required": [
-        "newValue"
-    ],
-    "properties": {
-        "newValue": {
-            "type": "string",
-            "title": "New IFRS value"
-        },
-    }
-};
